+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.1
+
+- Documentation and usage information
+- Executable is now ready
+
 ## 2.0.0
 
 - Actual initial version. Need to bump to override old library